requirements.txt
langchain
langchain-google-genai 
python-dotenv
streamlit
pydantic
langchain-groq
langchain-community 
langchain-core
pypdf
bs4